[python-package] Address PR review comments for Polars/Arrow support
- Remove POLARS_INSTALLED from compat.py (test-only, use pytest.importorskip instead)
- Replace module-level POLARS_INSTALLED/PYARROW_INSTALLED guards in test_polars.py
  with pytest.importorskip("pyarrow") / pytest.importorskip("polars")
- Move PyArrow DictionaryType column tests from test_polars.py to test_arrow.py
  where they belong (not Polars-specific)
- Add polars_DataFrame to all_x_types in test_sklearn.py so the sklearn
  parametrized data-type coverage tests include Polars DataFrames
- Add polars to .ci/pip-envs/requirements-latest.txt and
  .ci/conda-envs/ci-core.txt so CI actually runs the polars tests
- Add Python package development workflow to CONTRIBUTING.md (submodule
  init, cmake build, build-python.sh install --precompile, pytest steps)
