+++ .ci/conda-envs/ci-core.txt
@@ -23,6 +23,7 @@ joblib>=1.3.2
 matplotlib-base>=3.7.3
 numpy>=1.24.4
 pandas>2.0
+polars>=0.19
 pyarrow-core>=6.0
 python-graphviz>=0.20.3
 scikit-learn>=1.3.2
