+++ .ci/pip-envs/requirements-latest.txt
@@ -12,6 +12,7 @@ cffi>=2.0.0
 matplotlib>=3.11.0.dev0
 numpy>=2.5.0.dev0
 pandas>=3.1.0.dev0
+polars
 pyarrow>=24.0.0.dev0
 scikit-learn>=1.9.dev0
 scipy>=1.18.0.dev0
